requirements.txt
fastapi
uvicorn
python-multipart
passlib
bcrypt==3.2.2
python-jose[cryptography]
cryptography
psycopg2-binary
slowapi
aiofiles
cachetools
